backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-api-python-client
isodate
yt-dlp
ffmpeg-python
mutagen
requests
beautifulsoup4